+++ step-capstone/src/__tests__/SuggestionsTests.test.js
@@ -1,143 +1,8 @@
 import react from 'react'
 import '@testing-library/jest-dom'
-import {countCat, getScore, rank} from '../scripts/Suggestion'
+import {rank} from '../scripts/Suggestion'
 import {suggestionTestData} from "../testData/SuggestionTestData"
 
-/*
- * countCat testing
-*/
-//one item matches preferences
-test("Counts type with one preference match correctly", () => {
-  const userCat = ["familyFriendly", "outdoors"]
-  const place = { 
-    place_id: 0, 
-    place: {
-      types: ["tourist_attraction", "aquarium"] 
-    }
-  } 
-  expect(countCat(place, userCat)).toEqual(1)
-})
-
-test("Counts type with multiple preference matches correctly", () => {
-  const userCat = ["familyFriendly", "outdoors"]
-  const place = { 
-    place_id: 1, 
-    place: {
-      types: ["tourist_attraction", "zoo"] 
-    }
-  } 
-  expect(countCat(place, userCat)).toEqual(2);
-})
-
-test("Zero matches with user preferences", () => {
-  const userCat = ["nightlife"]
-  const place = { 
-    place_id: 0, 
-    place: {
-      types: ["tourist_attraction", "aquarium"] 
-    }
-  } 
-  expect(countCat(place, userCat)).toEqual(0);
-})
-
-test("Doesn’t double count category matches", () => {
-  const userCat = ["sightseeing"]
-  const place = { 
-    place_id: 0, 
-    place: {
-      types: ["tourist_attraction", "city_hall"] 
-    }
-  } 
-  
-  // tourist_attraction → sightseeing
-  // city_hall → sightseeing
-  expect(countCat(place, userCat)).toEqual(1);
-})
-
-test("Multiple types, multiple matches for each", () => {
-  const userCat = ["sightseeing", "outdoors", "historicalSites", "artsAndCulture"]
-  const place = { 
-    place_id: 0, 
-    place: {
-      types: ["tourist_attraction", "natural_feature", "church"] 
-    }
-  } 
-  // tourist_attraction → sightseeing, historicalSites
-  // natural_feature → outdoors
-  // church → sightseeing, historicalSites, artsAndCulture
-  expect(countCat(place, userCat)).toEqual(4)
-})
-
-/*
- * Score testing
- */
-
-const ZERO_PROMINENCE = { total: 1, index: 1 }
-
-test("Handles category score correctly", () => {
-  let score = getScore(2, 4, ZERO_PROMINENCE, 5, 1, 0);
-  expect(score).toEqual(52);
-})
-
-test("Matches all scores", () => {
-  let score = getScore(2, 2, ZERO_PROMINENCE, 5, 1, 0);
-  expect(score).toEqual(65);
-})
-
-test("Matches zero scores", () => {
-  let score = getScore(0, 2, ZERO_PROMINENCE, 5, 1, 0);
-  expect(score).toEqual(0);
-})
-
-test("Handles prominence correctly", () => {
-  let score = getScore(1, 1, { index: 5, total: 10 }, 5, 1, 0);
-  expect(score - 65).toEqual(3.75)
-})
-
-test("Handles #1 prominence correctly", () => {
-  let score = getScore(1, 1, { index: 0, total: 10 }, 5, 1, 0);
-  expect(score - 65).toEqual(15);
-})
-
-test("Handles rating correctly", () => {
-  let score = getScore(1, 1, ZERO_PROMINENCE, 5, 1, 3.5);
-  expect(score - 65).toEqual(7);
-})
-
-test("Handles 0 rating correctly", () => {
-  let score = getScore(1, 1, ZERO_PROMINENCE, 5, 1, 0);
-  expect(score - 65).toEqual(0);
-})
-
-test("Handles 5 rating correctly", () => {
-  let score = getScore(1, 1, ZERO_PROMINENCE, 5, 1, 5);
-  expect(score - 65).toEqual(10);
-})
-
-test("Handles no rating", () => {
-  let score = getScore(1, 1, ZERO_PROMINENCE, 5, 1, undefined);
-  expect(score - 65).toEqual(5);
-})
-
-test("Handles valid Budget correctly", () => {
-  let score = getScore(1, 1, ZERO_PROMINENCE, 1, 3, 0);
-  expect(score - 65).toEqual(10);
-})
-
-test("Handles null Budget correctly", () => {
-  let score = getScore(1, 1, ZERO_PROMINENCE, undefined, 3, 0);
-  expect(score - 65).toEqual(5);
-})
-
-// test("Too expensive", () => {
-//   let score = getScore()
-//   expect(score - 65).toEqual(0);
-// })
-
-test("Correctly calculates overall score", () => {
-  let score = getScore(3, 4, { index: 60, total: 120}, 2, 3, 4.4);
-  expect(score).toEqual(81.05);
-})
 
 /*
  * Overall Ranking Testing
@@ -229,7 +94,7 @@ test("Matching Type outranks Budget and rating", () => {
 })
 
 
-// make sure to filter out things we don’t want like lodging
+// TODO: make sure to filter out things we don’t want like lodging
 // test("Filter Invalid Types", () => {
 //   const config = suggestionTestData.prefMidBudgetFamilyEnt
 //   const testPlaces = new Map();
